requirements.txt
matplotlib==3.8.0
numpy==1.26.4
opencv_python==4.9.0.80
pandas==2.2.2
PyQt5==5.15.10
PyQt5_sip==12.13.0
pyserial==3.5
scipy==1.11.4
torch==2.2.2
